requirements.txt
gradio
newspaper3k
scikit-learn
nltk
networkx
joblib
lxml_html_clean
